renamed animation.h/c to animation_chain.h/c
