+++ DragonQuestino/animation_chain.h
