+++ DragonQuestino/player.c
@@ -42,16 +42,6 @@ void Player_Init( Player_t* player, TileMap_t* tileMap )
    player->stats.stopSpellResist = 0;
    player->stats.hurtResist = 0;
    player->stats.dodge = 1;
-
-   // MUFFINS: for testing
-   /*SPELL_SET_HASHEAL( player->spells );
-   SPELL_SET_HASSIZZ( player->spells );
-   SPELL_SET_HASMIDHEAL( player->spells );
-   SPELL_SET_HASZOOM( player->spells );
-   player->stats.maxHitPoints = 100;
-   player->stats.hitPoints = 80;
-   player->stats.maxMagicPoints = 100;
-   player->stats.magicPoints = 100;*/
 }
 
 uint8_t Player_GetLevelFromExperience( Player_t* player )
