+++ DragonQuestinoWinDev/DragonQuestinoWinDev/DragonQuestinoWinDev.vcxproj
@@ -19,7 +19,7 @@
     </ProjectConfiguration>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="..\..\DragonQuestino\animation.h" />
+    <ClInclude Include="..\..\DragonQuestino\animation_chain.h" />
     <ClInclude Include="..\..\DragonQuestino\battle.h" />
     <ClInclude Include="..\..\DragonQuestino\battle_stats.h" />
     <ClInclude Include="..\..\DragonQuestino\clock.h" />
@@ -45,7 +45,7 @@
     <ClInclude Include="win_pixel_buffer.h" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="..\..\DragonQuestino\animation.c" />
+    <ClCompile Include="..\..\DragonQuestino\animation_chain.c" />
     <ClCompile Include="..\..\DragonQuestino\battle.c" />
     <ClCompile Include="..\..\DragonQuestino\enemy.c" />
     <ClCompile Include="..\..\DragonQuestino\game.c" />
